Move media modules to actions dir
