+++ lib/fulcrum.rb
@@ -13,8 +13,8 @@ require 'fulcrum/classification_set'
 require 'fulcrum/form'
 require 'fulcrum/membership'
 require 'fulcrum/photo'
-require 'fulcrum/media_versions'
-require 'fulcrum/gps_track'
+require 'fulcrum/actions/media_versions'
+require 'fulcrum/actions/gps_track'
 require 'fulcrum/video'
 require 'fulcrum/audio'
 require 'fulcrum/signature'
