+++ lib/fulcrum/actions/gps_track.rb
