+++ lib/fulcrum/actions/media_versions.rb
